process tools designed as  .c and .h files
